backend/requirements.txt
flask==3.0.0
google-genai==0.2.2
python-dotenv==1.0.0
werkzeug==3.0.1